requirements.txt
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Flask==2.0.2
flask-swagger-ui==3.36.0
itsdangerous==2.0.1
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
pycparser==2.21
PyJWT==2.3.0
Werkzeug==2.0.2
